+++ _sass/layout/_l-header.scss
@@ -2,14 +2,28 @@ $header-height-desktop: rem(95px);
 $header-height: rem(70px);
 
 .l-header {
-  position: relative;
-  height: $header-height;
-
   display: flex;
   align-items: stretch;
   justify-content: space-between;
+  transition: transform 0.5s ease;
+  position: fixed;
+  top: -70px;
+  left: 0;
+  right: 0;
+  height: $header-height;
+  z-index: 20;
+  background-color: $white;
 
-  @media screen and (min-width: 768px) {
+  &.hide {
+    transform: translatey(0);
+  }
+
+  &.show {
+    transform: translatey(70px);
+  }
+
+  @media screen and (min-width: $screen-m) {
+    position: relative;
     height: rem(80px);
   }
 
@@ -43,43 +57,11 @@ $header-height: rem(70px);
     justify-content: flex-end;
 
     @media screen and (min-width: $screen-m) {
-      // display: flex;
       width: 100%;
-    }
-
-    .modal-background {
-      display: none;
-    }
-
-    &.is-open {
-      // .modal-background {
-      //   display: block;
-      //   transition-delay: .5s;
-      // }
-
-      .menus-wrapper  {
-        // transform: translatex(0);
-        // transition: all .3s ease-in .1s;
-
-        // .link:first-child {
-        //   a {
-        //     padding: rem(10px) rem(40px) rem(20px) 0;
-        //   }
-        // }
-      }
     }
 
     .menus-wrapper {
       display: none;
-      // position: fixed;
-      // top: 0;
-      // bottom: 0;
-      // right: 0;
-      // z-index: 100;
-      // width: 90%;
-      // background-color: $bg-color-4;
-      // transition: all .3s ease-in .1s;
-      // transform: translatex(100%);
 
       @media screen and (min-width: $screen-m) {
         width: 100%;
@@ -91,14 +73,6 @@ $header-height: rem(70px);
     }
 
     .scroll-wrapper {
-      // position: absolute;
-      // top: rem(60px);
-      // bottom: rem(60px);
-      // left: 0;
-      // right: 0;
-      // padding: 0 rem(40px);
-      // overflow-y: scroll;
-
       @media screen and (min-width: $screen-m) {
         position: static;
         display: flex;
@@ -130,30 +104,6 @@ $header-height: rem(70px);
       @media screen and (min-width: $screen-l) {
         height: $header-height-desktop;
       }
-    }
-
-    // .btn-close {
-    //   @media screen and (min-width: $screen-m) {
-    //     display: none;
-    //   }
-    // }
-  }
-
-  &.is-retractable {
-    transition: transform 0.5s ease;
-    position: fixed;
-    top: -70px;
-    left: 0;
-    right: 0;
-    z-index: 20;
-    background-color: $white;
-
-    &.hide {
-      transform: translatey(0);
-    }
-
-    &.show {
-      transform: translatey(70px);
     }
   }
 }
