+++ js/src/routers/blog.js
@@ -2,7 +2,7 @@ var Backbone = require('backbone');
 
 var ViewManager = require('../lib/view_manager.js'),
   BlogView = require('../views/blog/blog.js'),
-  WrapperHeaderView = require('../views/common/wrapper_header_view.js');
+  MobileMenuView = require('../views/common/mobile_menu_view.js');
 
 var Router = Backbone.Router.extend({
 
@@ -12,7 +12,7 @@ var Router = Backbone.Router.extend({
 
   initialize: function(options) {
     new ViewManager({ $el: options.$el });
-    new WrapperHeaderView();
+    new MobileMenuView();
   },
 
   blog: function() {
