+++ js/src/routers/indicators.js
@@ -3,7 +3,7 @@ var Backbone = require('backbone');
 var ViewManager = require('../lib/view_manager.js'),
   IndicatorsView = require('../views/indicators/indicators.js'),
   IndicatorView = require('../views/indicators/indicator.js'),
-  WrapperHeaderView = require('../views/common/wrapper_header_view.js');
+  MobileMenuView = require('../views/common/mobile_menu_view.js');
 
 var Router = Backbone.Router.extend({
 
@@ -17,7 +17,7 @@ var Router = Backbone.Router.extend({
 
     this._setListeners();
 
-    new WrapperHeaderView();
+    new MobileMenuView();
   },
 
   _setListeners: function() {
