+++ js/src/routers/welcome.js
@@ -2,7 +2,7 @@ var Backbone = require('backbone');
 
 var ViewManager = require('../lib/view_manager.js'),
   WelcomeView = require('../views/welcome/welcome.js'),
-  WrapperHeaderView = require('../views/common/wrapper_header_view.js');
+  MobileMenuView = require('../views/common/mobile_menu_view.js');
 
 var Router = Backbone.Router.extend({
 
@@ -16,7 +16,7 @@ var Router = Backbone.Router.extend({
 
   welcome: function() {
     new WelcomeView();
-    new WrapperHeaderView();
+    new MobileMenuView();
   }
 
 });
