+++ js/src/templates/common/mobile_menu_tpl.hbs
@@ -4,11 +4,11 @@
   <div class="scroll-wrapper">
     <nav class="m-site-navigation -header">
       <ul>
-        <li class='link {% if page.url == "/countries.html" %} -current {% endif %}'><a href="{{ site.baseurl }}/countries">Countries</a></li>
-        <li class='link {% if page.url == "/indicators.html" %} -current {% endif %}'><a href="{{ site.baseurl }}/indicators">Data</a></li>
-        <li class='link {% if page.url == "/compare.html" %} -current {% endif %}'><a href="{{ site.baseurl }}/compare">Compare</a></li>
-        <li class='link {% if page.url == "/about.html" %} -current {% endif %}'><a href="{{ site.baseurl }}/about">About</a></li>
-        <li class='link {% if page.isBlog or page.layout == "blog/post" %} -current {% endif %}'><a href="{{ site.baseurl }}/blog/">Blog</a></li>
+        <li class='link'><a href="/countries">Countries</a></li>
+        <li class='link'><a href="/indicators">Data</a></li>
+        <li class='link'><a href="/compare">Compare</a></li>
+        <li class='link'><a href="/about">About</a></li>
+        <li class='link'><a href="/blog/">Blog</a></li>
       </ul>
     </nav>
