+++ js/src/views/common/line_chart.js
@@ -1,8 +1,7 @@
 var _ = require('lodash'),
   $ = require('jquery'),
   Backbone = require('backbone'),
-  Handlebars = require('handlebars'),
-    
+  Handlebars = require('handlebars');
 
 var LineChart = Backbone.View.extend({
