+++ js/src/views/countries/country.js
@@ -1,6 +1,5 @@
 var _ = require('lodash'),
     Backbone = require('backbone'),
-    enquire = require('enquire.js'),
     Handlebars = require('handlebars');
 
 var Country = require('../../models/country.js'),
@@ -16,8 +15,7 @@ var CountryHeaderView = require('./country_header.js'),
     ModalWindowView = require('../common/infowindow_view.js')
     ShareWindowView = require('../common/share_window_view.js'),
     YearSelectorView = require('../common/year_selector.js'),
-    LegendView = require('../common/legend.js'),
-    RetractableMenuView = require('../common/retractable_menu_view.js');
+    LegendView = require('../common/legend.js');
 
 
 var template = Handlebars.compile(require('../../templates/countries/country.hbs'));
@@ -34,20 +32,6 @@ var CountryView = Backbone.View.extend({
     if (options.iso === undefined) {
       throw new Error('CountryView requires a Country ISO ID');
     }
-
-    enquire.register("screen and (max-width:769px)", {
-      match: _.bind(function(){
-        this.mobile = true;
-        this.initViews();
-      },this)
-    });
-
-    enquire.register("screen and (min-width:770px)", {
-      match: _.bind(function(){
-        this.mobile = false;
-        this.initViews();
-      },this)
-    });
 
     // Models
     this.infoWindowModel = new InfoWindowModel();
@@ -75,12 +59,6 @@ var CountryView = Backbone.View.extend({
 
     this._setListeners();
     this.initializeData();
-  },
-
-  initViews: function() {
-    if (this.mobile) {
-      new RetractableMenuView();
-    }
   },
 
   initializeData: function() {
