+++ js/src/views/indicators/indicator.js
@@ -6,18 +6,16 @@ var _ = require('lodash'),
 
 var Indicator = require('../../models/indicator.js');
 
+var FunctionHelper = require('../../helpers/functions.js');
+
 var Countries = require('../../collections/countries.js');
     YearsCollection = require('../../collections/years.js');
-
-var FunctionHelper = require('../../helpers/functions.js');
 
 var IndicatorHeaderView = require('./indicator_header.js'),
   IndicatorSelectorsToolbarView = require('./indicator_selectors_toolbar.js'),
   CountryListView = require('./country_list.js'),
   ShareWindowView = require('../common/share_window_view.js'),
   LegendView = require('../common/legend.js');
-
-var RetractableMenuView = require('../common/retractable_menu_view.js');
 
 var TextShortener = require('../common/text_shortener.js');
 
@@ -34,20 +32,6 @@ var IndicatorView = Backbone.View.extend({
 
   initialize: function(options) {
     options = options || {};
-
-    enquire.register("screen and (max-width:769px)", {
-      match: _.bind(function(){
-        this.mobile = true;
-        this.initViews();
-      },this)
-    });
-
-    enquire.register("screen and (min-width:770px)", {
-      match: _.bind(function(){
-        this.mobile = false;
-        this.initViews();
-      },this)
-    });
 
     window.indicatorId = this.id;
 
@@ -81,12 +65,6 @@ var IndicatorView = Backbone.View.extend({
     this.indicator.set({
       id: this.status.get('id')
     }).fetch();
-  },
-
-  initViews: function() {
-    if (this.mobile) {
-      new RetractableMenuView();
-    }
   },
 
   _setListeners: function() {
