+++ js/src/views/indicators/indicators.js
@@ -1,6 +1,5 @@
 var _ = require('lodash'),
   Backbone = require('backbone'),
-  enquire = require('enquire.js'),
   Handlebars = require('handlebars');
 
 var IndicatorConfigs = require('../../collections/indicator_configs.js');
@@ -9,37 +8,15 @@ var IndicatorList = require('./indicator_list.js');
 var template = Handlebars.compile(
   require('../../templates/indicators/indicators.hbs'));
 
-var RetractableMenuView = require('../common/retractable_menu_view.js');
-
 var IndicatorsView = Backbone.View.extend({
 
   initialize: function() {
-    enquire.register("screen and (max-width:769px)", {
-      match: _.bind(function(){
-        this.mobile = true;
-        this.initViews();
-      },this)
-    });
-
-    enquire.register("screen and (min-width:770px)", {
-      match: _.bind(function(){
-        this.mobile = false;
-        this.initViews();
-      },this)
-    });
-
     this.indicatorsData = null;
 
     // collections
     this.indicators = new IndicatorConfigs();
 
     this._getIndicators();
-  },
-
-  initViews: function() {
-    if (this.mobile) {
-      new RetractableMenuView();
-    }
   },
 
   _getIndicators: function() {
