+++ js/src/views/welcome/welcome.js
@@ -4,8 +4,7 @@ var $ = require('jquery'),
   Backbone = require('backbone');
 
 var SearchView = require('../common/search_view.js'),
-  SearchMobileView = require('../common/search_mobile_view.js'),
-  RetractableMenuView = require('../common/retractable_menu_view.js');
+  SearchMobileView = require('../common/search_mobile_view.js');
 
 var WelcomeView = Backbone.View.extend({
 
@@ -46,7 +45,6 @@ var WelcomeView = Backbone.View.extend({
       new SearchView({ el: $('.js--search') });
     } else {
       new SearchMobileView({ el: $('.js--search') });;
-      new RetractableMenuView();
     }
   },
